chore: extract initialize scope
