+++ src/Http/Scope/index.ts
@@ -1,2 +1,3 @@
 export type * from './types';
+export * from './initialize-scope';
 export * from './State';
