+++ src/Http/Scope/initialize-scope.ts
@@ -1,7 +1,12 @@
+import { type Application } from '@/application';
 import { type Next } from 'koa';
-import { type HttpResponse, type HttpScope } from '@/Http';
+import { type HttpResponse } from '../Response';
+import { type HttpScope } from './types';
 
-export async function extendResponse(scope: HttpScope, next: Next): Promise<void> {
+export async function initializeScope(scope: HttpScope, next: Next): Promise<void> {
+  const app = scope.app as unknown as Application;
+  app.scope = app.context;
+
   scope.response.setHeader = function (name: string, value: string | string[]): HttpResponse {
     scope.set(name, value);
